Assets module split up in progress
